Scrollable View for the Add User forms
