+++ onboard-jmuniz/components/users-list/add-users/add-user-pressable.tsx
@@ -1,0 +1,14 @@
+import styled from "styled-components/native";
+
+export const AddUserPressable = styled.TouchableHighlight`
+  background-color: gray;
+  position: absolute;
+  bottom: 20px;
+  right: 10%;
+  z-index: 1;
+  height: 60px;
+  width: 60px;
+  border-radius: 30px;
+  justify-content: center;
+  align-items: center;
+`;
