+++ onboard-jmuniz/components/users-list/add-users/pressable-container.tsx
@@ -1,0 +1,8 @@
+import { styled } from "styled-components/native";
+
+export const PressableContainer = styled.View`
+  height: 0%;
+  width: 100%;
+  margin: auto;
+  display: flex;
+`;
