+++ onboard-jmuniz/components/users-list/add-users/user-pressable-icon.tsx
@@ -1,0 +1,8 @@
+import styled from "styled-components/native";
+
+export const UserPressableIcon = styled.Text`
+  align-self: center;
+  font-size: 35px;
+  color: white;
+  font-weight: 600;
+`;
